Also move the parsers.
