+++ web/app/containers/reducers.js
@@ -3,7 +3,6 @@
 import {combineReducers} from 'redux'
 
 import app from './app/reducer'
-// import catalog from './catalog/reducer'
 import cart from './cart/reducer'
 import checkoutConfirmation from './checkout-confirmation/reducer'
 import checkoutPayment from './checkout-payment/reducer'
@@ -21,7 +20,6 @@ import plp from './plp/reducer'
 
 const rootReducer = combineReducers({
     app,
-    // catalog,
     cart,
     checkoutConfirmation,
     checkoutPayment,
